MDL-73679 core_courseformat: formats can override core templates
From Moodle 4.0 format plugins can override any courseformat output
class. With this commit format plugins will also be able to override
individual mustache templates if an alternative version is provided in
the course/format/PLUGIN/templates/courseformat folder.

The override template uses the same logic as the theme ones, this means
that the subfolder inside "templates/courseformat" should be the same as
the original template.