MDL-73679 example: More complete example
This extends the example to meet provide all of the required extension
points so that any part of the course will be updated, no matter how it
is updated.

There are three classes which are rendered at various times:
* content
* section
* cmitem

By providing renderer classes for each of these, and making the
templates extend one another with the standard values used, any part of
any of the templates or their inclusions can be overridden in whole, or
in part.