MDL-78181 javascript: centralise modal action protected dispatcher
This patch centralises the modal submitting logic (which was replicated
three times in the code) and also converted into a modal generic method
that other modal can use to provide alternative submission methods. This
is used for example in the radio modal to submit the form when the user
double clicks an option.