MDL-82349 core_coruseformat: new get_generic_section_name method
The string "secitonname" was an unnecessary coupling between
formats and other plugins. Now the generic name for a section
should be obtained using $format->get_generic_section_name.
This allow names to use an alternative string for nameing
sections. This is especially important for rare formats like
the frontpage one.